Remove default gap-set and alphabet settings, cleanup readme
Since we now have a notion of "unset variables", there is no need
for many default values. Having 'gap-set' set to [-1000 ... 1000]
was an ugly hack to simulate no gap-set restrictions, and 'alphabet'
didn't actually have a default value.

I have updated the readme to say '(none)' for variables which don't
have a default value set, and removed the "0 to disable" note for
max-depth and max-iterations, since they can now be unset properly.

The actual behavior when these are set to 0 is preserved, but if
profiling shows a speed improvement here, maybe one day I'll remove
that too.