Fix invalid symmetry assumption in coloring.c
When recursing on colorings, we assume that all colors are
essentially the same. So, for example, if colors 2, 3, and
4 are empty, the program will set the next number to color
2 and not bother with 3 and 4.

However, if the gap sets are different for different colors,
this assumption is invalid and the code broken.

This patch adds a has_symmetry flag to the coloring recursion
code, and only assumes symmetry if symmetric gapsets are used.