Add alternate internal representation for coloring.c
For some problems (specifically rainbow AP's), we need to represent
a coloring as a word on an alphabet; i.e.,

 [[ 1 2 5 ] [ 3 4 ]]      would be      00110

where our two colors are 0 and 1. This patch adds such a representation,
and maintains it alongside the old partition representation. (I have not
yet decided how to access this representation.)