Fix potential memory leak in iters-per-length.
In the constructor for the iters-per-length object, we allocate a
new stream object (as an output hook for the dump), then allocate
memory for the stream object itself.

If the memory allocation fails, we output a message and return NULL,
but we did not free the stream object. This commit fixes that.

(Caught by splint.)