Remove access to global state from subthreads
Subthreads now should only be dealing with their own private memory,
which will only be accessed by one other thread (the main one), and
only after they have died.

A consequence of this is that the max-length target now does not have
access to the actual object of maximum length. This will need to be
fixed.

But for now, we have a massive speedup.