Add `fork' target (for my use with WestGrid)
The `fork' target outputs a "search [space] [seed]" line, where
[space] is the current space (plus an extra 's'), [seed] is the
current object, and the output appears whenever we are at a depth
specified by the ``fork-depth'' variable.

So, for example, writing

  ## max-depth must be strictly greater than fork-depth
  set max-depth 5
  set fork-depth 4
  target fork

  filter no-double-3-aps
  search colorings

We will get a bunch of output of the form

  search colorings [[1, 3] [2] []]

which can then be plugged into new scripts and evaluated in their
own process. In a later commit I will add a ``quiet'' command to
eliminate the rest of the output, making it easy to do everything
through bash.