Add `quiet' command to toggle quiet mode
When you type 'search colorings', for example, RamseyScript
will dump a bunch of data about active targets, filters, etc,
which isn't really necessary in a scripted environment (since
all this data is available from the source script).

So now there is a 'quiet' command to suppress all this. Typing
'quiet' again will unsuppress it, and so on.