Fix fencepost error, bad memory access in dump/iters-per-length.c
There were some '<=' checks that should have been '<',
presumably because iters-per-length is supposed to act
like a 1-indexed array. But it wasn't done properly
and the result is valgrind complaining.

Now we just increment whatever the user sets as dump-depth,
and don't ever print the 0th index. This has the desired
behavior, e.g., when searching all sequences with gap-set
[1 2 3] the dump output is [ 3 9 27 81 ... ].