Remove Gtk
Since it is not really documented or used, it is probably just
confusing to have GUI code in RamseyScript.

We have not used the Gtk code in several months, and it is very
unlikely that it still works (if it ever did). Anyone interested
in reviving it should fork from the previous commit and merge in
later changes.